Update project with latest changes

+++ trading/src/main/java/com/chirag/trading/service/VerificationCodeImpl.java
@@ -1,3 +1,4 @@
+
 package com.chirag.trading.service;
 
 import com.chirag.trading.Model.User;
@@ -7,26 +8,36 @@ import com.chirag.trading.domain.VerificationType;
 import com.chirag.trading.utils.OtpUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.Optional;
 
 @Service
-public class VerificationCodeImpl implements VerificationCodeService{
+public class VerificationCodeImpl implements VerificationCodeService {
+
     @Autowired
     private VerificationCodeRepository verificationCodeRepository;
+
     @Override
     public VerificationCode sendVerificationCode(User user, VerificationType verificationType) {
-        VerificationCode verificationCode1=new VerificationCode();
-        verificationCode1.setOtp(OtpUtils.generateOtp());
-        verificationCode1.setVerificationType(verificationType);
-        verificationCode1.setUser(user);
-        return verificationCodeRepository.save(verificationCode1);
+        // Delete any existing verification code for this user
+        VerificationCode existingCode = verificationCodeRepository.findByUserId(user.getId());
+        if (existingCode != null) {
+            verificationCodeRepository.delete(existingCode);
+        }
+
+        // Create new verification code
+        VerificationCode verificationCode = new VerificationCode();
+        verificationCode.setOtp(OtpUtils.generateOtp());
+        verificationCode.setVerificationType(verificationType);
+        verificationCode.setUser(user);
+        verificationCode.setEmail(user.getEmail()); // Set email
+
+        return verificationCodeRepository.save(verificationCode);
     }
 
     @Override
     public VerificationCode getVerificationCodeById(Long id) throws Exception {
-        Optional<VerificationCode> verificationCode=verificationCodeRepository.findById(id);
-        if (verificationCode.isEmpty()){
+        Optional<VerificationCode> verificationCode = verificationCodeRepository.findById(id);
+        if (verificationCode.isEmpty()) {
             throw new Exception("Verification Code Not Found");
         }
         return verificationCode.get();
@@ -34,12 +45,13 @@ public class VerificationCodeImpl implements VerificationCodeService{
 
     @Override
     public VerificationCode getVerificationCodeByUser(Long userId) {
-
         return verificationCodeRepository.findByUserId(userId);
     }
 
     @Override
     public void deleteVerificationCodeById(VerificationCode verificationCode) {
-        verificationCodeRepository.delete(verificationCode);
+        if (verificationCode != null) {
+            verificationCodeRepository.delete(verificationCode);
+        }
     }
 }
